Add Google Gears support.
Fall back on Google Gears if supported and none of the other back ends are
available.  Also add a function, backendType, to return the name of the
back-end storage type as a string.

This is required to support some browsers, such as the Google G1 Android
mobile browser.  This can be tested with the development emulator running
version 1.6 of the OS.  The output of:

  document.write(navigator.userAgent + '<br>');
  document.write('Using back end ' + $.jStorage.backendType() + '<br>');

is (backslash inserted):

Mozilla/5.0 (Linux; U; Android 1.6; en-us; sdk Build/Donut) \
AppleWebKit/528.5+(KHTML, like Gecko) Version/3.1.2 Mobile Safari/525.20.1
Using back end Google Gears